updated delete interview action

+++ controllers/interview_controller.js
@@ -70,13 +70,18 @@ module.exports.create = async function (req, res) {
 //delete interview
 module.exports.delete = async function (req, res) {
   try {
-    const interview = Interview.findById(req.params.id);
-    console.log(interview);
-    interview.students.forEach(async (student) => {
-      await Student.findByIdAndUpdate(student.student, {
-        $pull: { interviews: { interview: req.params.id } },
+    const interview = await Interview.findById(req.params.id);
+    // console.log(interview);
+    // interview.students.forEach(async (student) => {
+    //   await Student.findByIdAndUpdate(student.student, {
+    //     $pull: { interviews: { interview: req.params.id } },
+    //   });
+    // });
+    if (interview.students.length > 0) {
+      return res.status(400).json({
+        message: "Interview has students",
       });
-    });
+    }
     await Interview.findByIdAndDelete(req.params.id);
     return res.status(200).json({
       message: "Interview deleted",
